Hello

+++ agency_portal/menukit/models.py
@@ -15,7 +15,7 @@ class Category(models.Model):
 class SubCategory(models.Model):
     name = models.CharField(max_length=100)
     category = models.ForeignKey(Category, on_delete=models.CASCADE, related_name='subcategories',verbose_name="Menu")
-    price = models.IntegerField(null=True)
+    price = models.FloatField(null=True)
     ingrediants = models.CharField(max_length=255, null=True,blank=True,default="")
     deleted = models.IntegerField(default=0)
